fix: better card layouts, fix wallet disconnect redirect (#32)
The layout of the cards in the billing section was bothering me, so
fixed that. On some pages that only make sense if your wallet is
connected, we want to redirect to the home page if you disconnect your
wallet. This logic was broken for the billing page, so I fixed that.

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>