Constants and chains packages (#52)
This PR adds two new workspace packages, `network-constants` which lists
all things like urls and contract addresses, etc, and `chains` which
defines all the wagmi chains and provides helper functions to work with
them. These were mostly taken from existing sdk code. This is useful
because we can use these packages now without depending on the entire
sdk and there is a single place now to update all things when a network
upgrade or reset happens.

Once this is merged, it will be nice and easy to make the portal app
configurable for which chain to run on.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>