Don't url encode entire path (#193)
The path usually includes `/` delimiters which shouldn't be encoded
because they are the url path structure that the objects api expects in
order to look up objects.

Ideally, we'd encode each component of the path, excluding the `/`, and
use that to build the URL, but the object api doesn't seem to support
URL ecdoded values, so it just returns not found in that case.
Basically, we need to fix the object api then we can make more
improvements here.

Closes #187

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>