Enable ts undefined for indexed access, deal with undefined values (#241)
Enabled a Typescript compiler option I think is very useful and safe...
When you access data by index, in an array for example, the result can
possibly be undefined. This is more correct, and the standard used in
the rest of the repo.

IMPORTANT: Logic changes here to deal with possibly undefined values.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>