Rafael/wallet auth (#228)
https://github.com/recallnet/js-recall/issues/225

Just added some basic auth with the tools requested. Currently the
coinbaseWallet config already handles the wallet creation with the
window used for signing.

I think next steps with this would be to add JWT and validate the nonce
properly. For this we will need the db and db schema. Also added an
endpoint to generate a nonce.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>