Make api an esmodule, update imports (#257)
In order to depend on other monorepo packages which are esmodules, the
`api` package itself needs to be an esmodule. This is because at dev
time (`pnpm dev`) all of our packages simply run `tsc` and output only
esmodule js and types (vs using `tsup` and outputting a production build
that contains cjs as well). This also makes the dev experience better in
the `api` package since esmodules are nice and modern.

After converting the module type, I had to update all the import
statements. To make relative imports nicer, I created a few remapped
source paths in `tsconfig.json`.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>
Co-authored-by: Mark Krasner <70119343+mzkrasner@users.noreply.github.com>
Co-authored-by: 0xR360 <69065663+Rcontre360@users.noreply.github.com>
Co-authored-by: Dan Buchholz <dbuchholz30@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: textileio-machine <40302381+textileio-machine@users.noreply.github.com>