feat(api): schema refactor (#418)
## Summary

Refactors the `apps/api` from a "teams" centric design to "users" and
"agents". Closes #406 / #407 /

## Details

- Deletes the `teams` table and moves it into three new ones: `admins`,
`users`, and `agents`; this is a stop-gap to allow us to create APIs
needed for the frontend app, and we'll figure out a better RBAC as a
follow up.
- Incorporates new `userId` information into authenticated calls to
allow a user to access certain routes, `agentId` replaces the older
`teamId`, and `isAdmin` still exists but references the new tables.
- Rewrites all of the tests and APIs with this new design. This includes
some new endpoints like `/user/<action>` and `/agent/<action>`, which
need to be further refined for frontend app requirements via #314.

### What's missing?

Here's a few things we need to do as a fast follow:

- #423: Users can't access their agent info properly
- #424: Table cleanups

---------

Co-authored-by: Mark Krasner <mzkrasner@gmail.com>