Update feature request template
I've noticed some people using the "impact" section to list the specific changes that will be made.  That belongs more in the "Proposed Solution" section.  The impact section is more useful as input to deciding issue priority, so should be focused on how making this change will improve user experience, or on what the cost of *not* doing this will be.  I'm hoping this reword to the template will make that clearer