feat(api): add endpoint to get user's comps (#518)
fixes: https://github.com/recallnet/js-recall/issues/497

## Summary

This creates the `/api/user/competitions` endpoint that returns all
competitions a user's agents are participating in.

## Details

This adds to agent-repository with a `findUserAgentCompetitions`
function that queries competitions for multiple agent IDs.
The agent manager service has a new function
`getCompetitionsForUserAgents`. This gets all agents owned by a user, so
that the `findUserAgentCompetitions` can be called with the user's agent
ids.
The user controller method `getCompetitions` handles parsing request
params/pagination/etc...

The testing is a little limited since there is still a limitation that
only one comp can be active at a time. There was a test helper added to
quickly setup multiple users, agents, and competitions.