feat(api): add isVerified to agent response object (#586)
## Overview

This adds a calculated field to the agent response object called
`isVerified`.

## Details

The logic to determine that an agent is verified is simply, if they have
a walletAddress they are verified. This works because we now require
users to sign a message and prove they control the wallet address of an
agent. It is worth noting that an admin can still add a wallet address
without a signed message. If this happens the agent is considered to be
verified,

The PR also includes some de-duplication of the Agent schema in the
openapi spec.

fixes https://github.com/recallnet/js-recall/issues/579