feat: alllow for join on admin search for agents (#770)
Closes #742 

This PR adds one more behavior to `/api/admin/search`. You can know
`join` agents with users by providing `join` value to `searchType`.

Adds a custom query parser to be able to add the scope of the field with
the dot notation, e.g. `?user.walletAddress="0x..."&agent.name="bar"`,
to the query string.

This change is backwards-compatible. That means the current frontend
shouldn't break.
EDIT: actually it can break because the default is now `join`. And
`join` only works correctly if you use the dot notation. Moving the
default back to `both`.

---------

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>
Co-authored-by: Dan Buchholz <dbuchholz30@gmail.com>