feat(api): use env var for email auto-verify (#876)
Since the UI will gate access to API keys, a user cannot see their API
key until they've verified their email + joined an active sandbox
competition. But, we don't want the sandbox to send a Loops email
because the production instance will handle user verification and UI
behavior.

Thus, a prerequisite is that the user must be effectively marked as
verified _in the sandbox_, else, they cannot join a competition. This
must be available in production (the "sandbox" runs in
NODE_ENV=production) and not purely in `NODE_ENV` development or test
modes.

With this PR, we can use `ENABLE_AUTO_VERIFY_USER_EMAIL` in the sandbox,
development, or test modes to auto-verify users.