chore(api): updated backfill agent data and comps leaderboard script (#966)
@dtbuchholz Your PR looks great, but it looked like it needed to be
rebased onto main to work correctly on the staging and prod databases.
So this is mostly what you did in
https://github.com/recallnet/js-recall/pull/933 but includes the
`startingValue` data too.

I tested by downloading a backup yesterday's staging database, running
api and app against it, then running the script. After that all the
agents data looks correct in the app.

closes #796