fix(api): remove active comp cache (#1013)
the "auto end competition" logic fails to trigger due to faulty "active
competition" cache. we observed:
1. start/end comp 1 => manually triggered comp end via admin API
2. start/end comp 2 => manually triggered comp end via admin API
3. start comp 3 with an end date => use auto-end logic => fails to end
(and never ends it on subsequent checks)

the logs state the "current" active competition is the comp from step
(1) and not the previous comp in (2). thus, when you try to auto-end
(3), it fails.

---

furthermore, the problem with the current caching is that it auto-loads
the current active competition upon service startup and pre-sets the
cache with this value. since we're pulling the scheduler _out of the
API_ and into its own cron, this adds a risk point for things to get out
of sync. hence, why it should be removed (and/or we make this more
sophisticated in the future).