fix(api): refactoring the e2e test startup and shutdown (#1005)
fixes: #881 
This has turned into a large refactor. There's a few categories of
changes.

## duplicated setup and teardown logic

I'm not really sure how our setup and teardown logic was generated, but
there was quite a bit of duplicated logic. There were also a few
comments mentioning Jest, which we have never used in this package
afaik. I'm assume there was some LLM hallucinations somewhere in the
past that made it onto `main`. Another thing I found that is a little
tricky if you aren't familiar with Vitest is that the config is set so
that `e2e/tests/utils/test-setup.ts` file is run for every file, and the
`e2e/setup.ts` file is only run globally once. I think this had resulted
in some unnecessary duplication of setup and teardown.

## incorrect handling of process shutdown

The e2e tests are run via Vitest. During startup for the tests, the
`server.ts` file uses the Node.js child_process spawn function to start
the api server in a sub-process. The command to start that server is
`npx tsx src/index.ts`. [`tsx`](https://www.npmjs.com/package/tsx) is
very useful in that it allows running typescript as if it was node.js
flavored javascript. The way this works involves starting 4 different
processes that all need to be managed by `tsx`, and although shutting
down the `tsx` command with `ctrl+c` works when you run the command in
your shell, it does not work how you might expect when trying to do
`process.kill(pid)` for the sub-process started by the `server.ts` file.
The solution I found is to use the
[`tree-kill`](https://www.npmjs.com/package/tree-kill) to handle killing
all of the processes.
I also removed the `detached: true` option from the process.spawn call
since we want to be able to do `ctrl+c` while the tests are running and
actually kill everything. (which was not possible before)

## confusion on in-memory availability of services during test running.

As mentioned above the actual server under test is in a sub-process, so
we can't access things like `setInterval` and in-memory caching. There
were a few places in the tests and setup files where we were trying to
interact with Services classes that were instantiated in the test runner
process, which doesn't do really do anything.