refactor: rewards should point to competitions and not epochs (#1107)
# Context

Closes
[APP-136](https://linear.app/recall-labs/issue/APP-136/rewards-database-schema-changes)

We already had some rewards tables from [previous work
](https://github.com/recallnet/js-recall/pull/436). But those tables
were pointing to `epochs` table. As of latest spec, we are implementing
rewards by competition for TGE.

This PR:
- Removes epoch from `rewards`, `rewards_tree` and `rewards_roots`
table, and add `competition_id`. This is a breaking change, but those
tables are not used. So, not a big deal.
- Refactors `rewards-repository`, `rewards.service`, `rewards.test.ts`
and scripts to accommodate that change

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>