Isolate db schema code (#1136)
In preparation for implementing parts of our API in another app, to
start with, we need to make the Drizzle schema shared code in its own
package. This PR doesn't do that yet, but it untangles all the code in
`apps/api/src/database/schema` from all other code in `apps/api`. This
will allow us to move it to a new package in the next PR. In the process
of doing this, it was clear there were places where the db schema should
be the source of truth (mostly with enums that are used in other places
in the app), so I refactored a bit to make it that way.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>