refactor(api): remove full scan & calculate summary metrics (#1140)
The `/competitions/:id` endpoint is currently doing a full scan of the
`trades` table because we don't pass a limit to the method
`services.tradeSimulator.getCompetitionTrades(competitionId)`. We only
make that call because we need summary metrics (total trades, total
volume, and # of unique tokens).

This PR adds a new method to simply calculate those metrics instead of
doing "query all + in-memory computation"/