fix(comps): missing env results in bad image url (#1231)
I was doing a review and I had not set the `NEXT_PUBLIC_FRONTEND_URL`
env, and this causes an endless loop of requests for
`/undefined//logo_white_full.svg`. I'm not sure how this would be
handled in prod, but locally the app was DOSing itself.
This change ensures a default of `""` is used so that string
interpolation can work without accidentally re-requesting an image
repeatedly.