fix(comps): retrieving agent api keys (#1224)
closes
[APP-419](https://linear.app/recall-labs/issue/APP-419/users-cannot-get-api-key-in-comps-ui-due-to-sandboxprod-db-mismatch).
a few fixes:
1. legacy users in the sandbox db might _only_ have a wallet address
- before: when we look up the agent API key, we were using the admin
search API for user + agent combinations (`user.walletAddress` +
`agent.handle`). this can lead to problems in case the prod<>sandbox dbs
get out of sync with user profile information.
- now: since agent handles are globally unique, we can be more lenient
and simply fetch the sandbox agent API key based on their handles. our
existing logic better handles updates to the sandbox if the agent handle
changes.
2. agent didn't exist
- before: somehow, the frontend didnt trigger an agent to get created.
when a user tried to get the sandbox API key, there was nothing to
fetch.
- now: we will create a sandbox agent if the agent handle is not found.
3. cant get prod keys, even though logged in
- before: our old logic had wonky email verification that gated both
prod and sandbox keys.
- now: production API keys are always shown to the user. sandbox API
keys are hidden, but only because the user needs to manually "join
competition". (note: we can probably refactor this and auto-join agents
into the sandbox competition.)