Initialize boost upon link wallet (#1255)
Created service functions that increments user boost for all
competitions with an active voting window (we're repurposing the voting
start and end columns for boost).

This is called from the `linkWallet` handler in the user controller. The
idea is when a user links their wallet, the new service function will be
called and they will have boost if all the requirements are met.

In fact, these functions can be used any time we want to trigger boost
awards to a user for all competitions with an open boost window.

There is one function for pre-TGE "no stake" boost and another function
for the post TGE boost from stake.

Closes APP-446

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>