Zod-based config (#1345)
When wiring up all the services and repositories as part of an upcoming
PR, dealing with env vars and config objects to pass into services was
getting unwieldy. I've also always wanted to have typesafe env vars to
use, so went ahead an build something. It covers all the configuration I
currently need for wiring up services, and I added a couple more comps
app specific things to it just to see how it works. If we like this
solution, we can add more to it.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>