feat: updates activate boost logic (#1394)
# Context

Changes the logic of the "Start boosting" button that allocates some
boosts to the user, with a feature flag. We now should consider the
stakes to determine how much boost to allocate.

Closes
[APP-599](https://linear.app/recall-labs/issue/APP-599/update-activate-boost-logic-for-tge)

This was manually tested by inserting stakes in the database, and
manually changing the code to activate one logic or the other.

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>