fix(comps): update env and monkey-patch rpc link (#1424)
I noticed that the example envs for the comps app doesn't have all the
required `POSTGRES_URL` var, and the privy auth button is spamming the
console with errors.
I might be missing some new config, but if not this change works around
the issue until we have a fix.