Wallet state fixes + txn state fixes (#1430)
We have to deal with the fact that privy users may authenticate, but not
connect/link a wallet. This PR properly keeps the ui in an appropriate
state and also detects if the user connects with the wrong wallet or on
the wrong chain.

Also fixes our txn success listeners to avoid race conditions where the
eth rpc provider tells us a txn is successful, but then returns contract
state that is not yet updated.

---------

Co-authored-by: Ramiro Vazquez <ramirovazquez@gmail.com>
Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>