cookie consent banner (#1431)
This implements a cookie consent modal. It is somewhat complex because
we need to restructure that ordering of provider components in order to
ensure cookies are not set incorrectly.

If you are testing you can open dev tools and delete all local storage
and cookies in the Application tab, or potentially leverage incognito,
or both.

---------

Co-authored-by: Dan Buchholz <dbuchholz30@gmail.com>