fix: Calmar ratio chart was double paginating (#1471)
fixes:
https://linear.app/recall-labs/issue/APP-665/fix-negative-calmar-values-not-showing-in-comp-chart

This issue stems from the fact that the api response handle pagination
for us (which is good), but the client is trying to also apply
pagination. This means when page 1 is paginated the chart takes the
first 10 agents, but when page 2 is paginated the chart takes the
11th-20th agents, but the array only has 10 agents. So, nothing is
displayed.
 
There is a separate but related issue(or maybe just confusing state)
where the 21st, i.e. last, place agent has a calmar ratio of null. This
results in the chart showing that there will be 3 pages, but after the
client filters the response there is only actually 2 pages.
I'm gonna look into that separately.