chore(api,db): remove dup idx per supabase advisor (#1478)
## Overview

The supabase dashboard has 4 warnings that we have duplicated indexes.
This PR updates Drizzle definitions when possible.

## Details

Because some of the supabase duplicated indexes were not added via
Drizzle, so we will have to manage them manually through the dashboard.
When we are ready to deploy this to production, before we run the
migration we will need to go to the supabase dashboard and manually do
the following:

- Delete the `idx_competition_agents_comp_status` index. LInk:
https://supabase.com/dashboard/project/fefmaihjhhkpzsnetesg/database/indexes?schema=public&table=agents
- Table trading_comps.portfolio_snapshots has two identical indexes, and
a third nearly identical index.
idx_portfolio_snapshots_comp_agent_ts, and
idx_portfolio_snapshots_optimal should be deleted. Note the migrations
in this PR modify the third index to the correct value.
Link:
https://supabase.com/dashboard/project/fefmaihjhhkpzsnetesg/database/indexes?schema=trading_comps&table=portfolio_snapshots

Part of:
https://linear.app/recall-labs/issue/APP-511/review-supabase-query-performance-dashboard-and-identify-potential