fix(comps): misc wallet cleanups (#1448)
## Summary

Minor nits with wallet connection

<details><summary>Legacy PR details</summary>
<p>

## Summary

Fixes endless `WrongWalletModal` popups if you've ever connected
phantom.

## Details

the `useWallets` hook will always include _any_ wallet that has been
linked to the app frontend. for example, a user has two accounts:
account A that uses metamask, and B that uses phantom. that means the
user has technically connected **both** wallets to the site. this is an
edge case but can certainly be annoying because it makes it impossible
to close the `WrongWalletModal`.

but, since wallet connections are subsequent to user privy logins, it
looks like privy will include a `linked` boolean value for every
connected (i.e., user owned / browser) wallet. for account A (metamask),
the following is part of the `wallets` response from the `useWallets`
hook. it shows `"linked": false`, which means we can use this filter out
"incorrect" injected wallets:
```json
{
    "type": "ethereum",
    "address": "0xEc7A69f205C79Dca5B9fD7CEfba5663Fe1e53Af0",
    "chainId": "eip155:84532",
    "meta": {
        "name": "Phantom",
       
        "id": "app.phantom"
    },
    "imported": false,
    "connectedAt": 1760411955541,
    "walletClientType": "phantom",
    "connectorType": "injected",
    "linked": false
}
```
or—if you were to log in with account B, it'll actually _skip_ the
connect wallet flow and auto-connect the phantom wallet. this is "okay"
because in this case, it's the correct wallet for the user, so they dont
run into `WrongWalletModal` issues. but, the account B flow gives you an
idea of how phantom vs others work in this hook.

</p>
</details>

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>