perf(db): replace 3s mean ex time query (#1484)
The database query perf dashboard shows that one of the agentBoost
queries has a mean time consumed of 2,034ms and is maxing out at
23,134ms. As best I can tell this is the result of locking the row
during insert, which means only one person at a time can boost a given
agent.

Here is the currently deployed query:
```
select
  "id",
  "agent_id",
  "competition_id",
  "total",
  "updated_at",
  "created_at"
from
  "agent_boost_totals"
where
  (
    "agent_boost_totals"."agent_id" = $1
    and "agent_boost_totals"."competition_id" = $2
  )
limit
  $3
for update
```

The results of the above query are only used if the idempotent key
triggers a noop, so I've removed the lock and only do the query when
needed.

Part of:
https://linear.app/recall-labs/issue/APP-511/review-supabase-query-performance-dashboard-and-identify-potential