fix(db): incorrect best placement
The best placement data is derived from 'getAgentRankingsInCompetitions'. This method queries the portfolio snaphots table for ended comps, but instead, it should use the cached competitions leaderbaord table. The bug technically happens because it's possible for a snapshot to come after the comp ended—at least, this is true for older comps. The cached copmetition leaderboards table is always correct for ended comps, so we should use it.