feat(api): enhance `flatParse` & make errors more descriptive (#1496)
The current `flatParse` method provides a vague error context, so it's
difficult to know what fields are invalid. The method is only used in
the admin API and leads to challenges with debugging (e.g., random
500s). This PR (partially stolen from
https://github.com/recallnet/js-recall/pull/870) changes the error
message to be more helpful (e.g., it includes the actualy violated
field) and also directly throws an `ApiError` with a 400. For example,
instead of `Invalid request: expected string, received undefined`, we'll
see the actual field(s) in the error: `id (expected string, received
undefined); email (invalid email address); ...`.