feat: automatically propose safe tx for allocating rewards (#1524)
# Summary

At the time we calculate and allocate rewards, we are creating a Safe
transaction that posts the merkle root in the Rewards Contract.
Eliminating the manual step of creating the transaction. But signers
still need to sign.

Closes
[ENG-771](https://linear.app/recall-labs/issue/ENG-771/propose-a-safe-transaction-for-allocate-rewards-after-rewards-are)

# Details
- This PR creates an interface called `RewardsAllocator` and 3
implementations:
- `NoopRewardsAllocator`: does nothing, and returns and "empty" tx
`0x0000000000000000000000000000000000000000`.
- `ExternallyOwnedAccountAllocator`: it's a normal EOA wallet doing the
smart contract call. It's useful for testing
- `SafeTransactionProposer`: proposes a transaction for a Safe Multisig
Account. This is the one we'll set in production to avoid manual steps.
- The env decides which will implementation is picked
- The rewards service that allocates rewards make use of that interface
without knowledge of the implementation
- We make `startTimestamp` optional with a default of
`end_competition_date` + `1 hour`. This is important when we fully
automate the process.

# Testing
This was manually tested with a personal Safe Account and data for
current running prod competition

# Todos
- [x] Adjust env on `competition-app-deployments` ( See
https://github.com/recallnet/competition-app-deployments/pull/37)

---------

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>