Comps lint + GDPR cleanup (#1519)
The comps app was using `next lint` to run lint, and that program only
lints files in the `app` directory. This was ignoring all of the other
code in the app, which is most of the code. This linting behavior
couldn't be modified via config, and in newer versions of NextJS, they
actually drop `next lint` in favor of direct `eslint` usage. We should
upgrade NextJS soon, but for now I just moved us to using `eslint`.

This change uncovered all sorts of problems, all of which are fixed in
the PR.

The biggest of those problems was the wagmi and privy mocks that were
used if a user didn't accept cookies. Once linting was on for those
files, it exposed React usage problems and compilation problems. This
got me to evaluating if totally disabling login and wallet interactions
is actually required for GDPR. It's not. In fact it's against GDPR's
rules to disable access to your site if cookies aren't accepted. Privy
is now under the "necessary" category of cookies, and wagmi isn't at all
related to cookies or GDPR.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>