Add indexing for the claim transaction in conviction claiming. (#1514)
In hopes of keeping the PR reasonable in size, this only includes the
updates to the Database to store claim data and the change to the
indexer to collect the data.

Future PRs will handle enabling updates to the airdrop contract and
monthly claims.

ENV change: The `INDEXING_START_BLOCK` Needs to be changed to
`INDEXING_EVENTS_START_BLOCK`