fix: filter proofs if rewards were not allocated (#1575)
# Summary


Closes
[ENG-952](https://linear.app/recall-labs/issue/ENG-952/correctly-handle-claim-button-if-safe-tx-is-not-approved)

# Details

The UI rewards claim button uses the following code to decide to show or
not:

```
const rewards = await context.rewardsService.getRewardsWithProofs(
    context.user.walletAddress,
);
 ```

`getRewardsWithProofs` is returning event if merkle root hasn't been published yet (waiting for signers approval) but rewards calculated. The UI assumes the returned proofs has a root already published.

The fix is to only return proofs if merkle root has been published. And we do that by checking if `tx` is set on `rewards_roots` table. That value is updated when we capture `AllocationAdded` event.

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>