Port scripts and db schema from claim portal to enable secondary conviction airdrops (#1556)
## Overview 
The goals here are to port the functionality from the claim portal app
to enable database and api changes for secondary airdrops for conviction
stakers.

## Details
The rules around claiming and staking remain generally the same as the
original airdrop. The longer you stake for the more token you receive.
The determination of how much token you are eligible in the ongoing
airdrops is based on the amount you have staked on the ex-date for the
airdrop of the given month and the amount remaining in the conviction
pool (un-claimed stake bonuses). This gives a significant increase in
token allocations for folks who stake the full 12 months.

---------

Co-authored-by: Aaron Sutula <528969+asutula@users.noreply.github.com>
Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>