wip: cleanup the calculate next season script (#1601)
The current plan is to make future seasons line up with the staking
durations, i.e. 30 day intervals. The problem is that "1 month" is only
an estimate, so depending on the month in question a 30 day stake is
shorter than the month.

The changes here will make the connection between qualifying for
conviction rewards and staking duration less complicated to understand.
If a given airdrop stake is still locked at the open of a "season" that
stake is included in the calculation for the next reward.
There is still the "1 year" option which is 365 days, (equal to a year
until 2028). Choosing a year will give people an "extra" season, but
this is ok since they are the most committed users. Also, assuming
everyone realizes the advantage of staking the full year the pool will
be exhausted before the year is up.

---------

Co-authored-by: Aaron Sutula <asutula@users.noreply.github.com>