feat(db): drop rewards `address` column & use `walletAddress` (#1674)
## Summary

Removes `rewards.address` column and uses `rewards.walletAddress`.
Closes
[ENG-1093](https://linear.app/recall-labs/issue/ENG-1093/clean-up-nullable-wallet-address-vs-address).

## Details

The `rewards.address` column is a legacy column that existed prior to
our user wallet standardization refactor. This simply removes the column
and uses the string wallet address column, which is migrated to be NOT
NULL. (E.g., query the `rewards` table to validate this: `select * from
rewards where wallet_address is null` => no results.)